pki.Descriptor: moved marshalling from impl to main.
